#60 (#94): add page for pending offers in Manage Pending Offers
