+++ SwiftParcel.Web/frontend/src/pages/offers/offersUser.tsx
