+++ SwiftParcel.Web/frontend/src/pages/offers/pendingOffers.tsx
@@ -1,0 +1,5 @@
+import OffersOfficeWorker from "./offersOfficeWorker";
+
+export default function PendingOffers() {
+  return OffersOfficeWorker("pending-offers");
+}
